+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Release Notes
 
-## Release 0.15.0
+## Unreleased
+
+### Added
+
+- Automatic replace of `@t-raw` to `@t-out` in `.xml` files during forwardport
+- Added separate command for automatic module migrate `odoo-helper ci auto-migrate-modules`
+  that will try to automarically migrate module code to current odoo version.
+  Currently it can do only simple *find-and-replace* via regex.
+
+
+---
+
+## Release 0.15.0 (2022-10-25)
 
 ### Added